# Pool bufio.Writers in cacheproxy