Before:
```
BenchmarkWrite/digest128B_chunk32B-32         	     822	   1221847 ns/op	 4021282 B/op	     748 allocs/op
BenchmarkWrite/digest128B_chunk64B-32         	    1030	   1302124 ns/op	 4023154 B/op	     748 allocs/op
BenchmarkWrite/digest128B_chunk96B-32         	     838	   1414660 ns/op	 4019968 B/op	     746 allocs/op
BenchmarkWrite/digest128B_chunk128B-32        	     890	   1161842 ns/op	 4019163 B/op	     746 allocs/op
BenchmarkWrite/digest16KiB_chunk4KiB-32       	     964	   1135321 ns/op	 4075474 B/op	     742 allocs/op
BenchmarkWrite/digest16KiB_chunk8KiB-32       	    1150	   1029718 ns/op	 4063932 B/op	     739 allocs/op
BenchmarkWrite/digest16KiB_chunk12KiB-32      	    1123	    985243 ns/op	 4058460 B/op	     738 allocs/op
BenchmarkWrite/digest16KiB_chunk16KiB-32      	    1185	   1002412 ns/op	 4054743 B/op	     737 allocs/op
BenchmarkWrite/digest16MiB_chunk4MiB-32       	      43	  30729754 ns/op	83923037 B/op	    1997 allocs/op
BenchmarkWrite/digest16MiB_chunk8MiB-32       	      48	  29842846 ns/op	79692623 B/op	    2109 allocs/op
BenchmarkWrite/digest16MiB_chunk12MiB-32      	      44	  37456051 ns/op	92263632 B/op	    2073 allocs/op
BenchmarkWrite/digest16MiB_chunk16MiB-32      	      50	  33577102 ns/op	71292733 B/op	    1945 allocs/op
```

After:
```
BenchmarkWrite/digest128B_chunk32B-32         	    2270	    517817 ns/op	   87799 B/op	     734 allocs/op
BenchmarkWrite/digest128B_chunk64B-32         	    2427	    502233 ns/op	   94925 B/op	     734 allocs/op
BenchmarkWrite/digest128B_chunk96B-32         	    2355	    502015 ns/op	  100088 B/op	     735 allocs/op
BenchmarkWrite/digest128B_chunk128B-32        	    2389	    510164 ns/op	   99157 B/op	     734 allocs/op
BenchmarkWrite/digest16KiB_chunk4KiB-32       	    2174	    558046 ns/op	  181205 B/op	     735 allocs/op
BenchmarkWrite/digest16KiB_chunk8KiB-32       	    2164	    522837 ns/op	  150119 B/op	     735 allocs/op
BenchmarkWrite/digest16KiB_chunk12KiB-32      	    2071	    572209 ns/op	  157496 B/op	     735 allocs/op
BenchmarkWrite/digest16KiB_chunk16KiB-32      	    2020	    563854 ns/op	  142738 B/op	     734 allocs/op
BenchmarkWrite/digest16MiB_chunk4MiB-32       	      39	  26817378 ns/op	82198285 B/op	    2038 allocs/op
BenchmarkWrite/digest16MiB_chunk8MiB-32       	      49	  29385956 ns/op	77236391 B/op	    2027 allocs/op
BenchmarkWrite/digest16MiB_chunk12MiB-32      	      45	  36296783 ns/op	89501597 B/op	    1991 allocs/op
BenchmarkWrite/digest16MiB_chunk16MiB-32      	      44	  33560419 ns/op	68344776 B/op	    2022 allocs/op
```